backend/requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
Pillow
python-multipart
beautifulsoup4
requests
aiohttp
